# frontend/openapi.yaml
openapi: 3.0.0
info:
  title: Stark Byte Hub API
  version: 1.0.0
  description: Documentation de l’API du backend Stark Byte Hub
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /logs/robot:
    get:
      summary: Liste les logs des robots
      tags:
        - Logs
      security:
        - cookieAuth: []
    delete:
      summary: Supprime tous les logs robot
      tags:
        - Logs
      security:
        - cookieAuth: []
  /logs/robot/{id}:
    get:
      summary: Récupère un log robot par ID
      tags:
        - Logs
      security:
        - cookieAuth: []
    delete:
      summary: Supprime un log robot par ID
      tags:
        - Logs
      security:
        - cookieAuth: []
...
  /robots:
    get:
      summary: Liste des robots liés à l'utilisateur connecté
      tags:
        - Robots
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des robots liés à l'utilisateur
    post:
      summary: Créer un nouveau robot
      tags:
        - Robots
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - serialNumber
                - linkKey
                - model
                - firmware
                - color
                - controllable
              properties:
                serialNumber:
                  type: string
                linkKey:
                  type: string
                model:
                  type: string
                firmware:
                  type: string
                color:
                  type: string
                controllable:
                  type: boolean
      responses:
        "201":
          description: Robot créé
        "403":
          description: Accès refusé
  /robots/all:
    get:
      summary: Liste de tous les robots en base
      tags:
        - Robots
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste complète des robots
        "403":
          description: Permission refusée
  /robots/link:
    post:
      summary: Lier un robot existant à l'utilisateur
      tags:
        - Robots
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - serialNumber
                - linkKey
                - name
              properties:
                serialNumber:
                  type: string
                  example: SBA-MX-V1-24040189
                linkKey:
                  type: string
                  example: ABC123XYZ
                name:
                  type: string
                  example: "Arachnobot #1"
      responses:
        "200":
          description: Robot lié avec succès
        "400":
          description: Clé ou numéro de série invalide
  /robots/import:
    post:
      summary: Importer plusieurs robots depuis un fichier CSV
      tags:
        - Robots
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Robots importés
        "400":
          description: Format de fichier invalide
        "403":
          description: Permission refusée
  /robots/{id}:
    get:
      summary: Récupère un robot par ID
      tags:
        - Robots
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Robot trouvé
        "403":
          description: Accès refusé
        "404":
          description: Robot introuvable
    put:
      summary: Met à jour un robot existant
      tags:
        - Robots
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                firmware:
                  type: string
                color:
                  type: string
                controllable:
                  type: boolean
      responses:
        "200":
          description: Robot mis à jour
        "403":
          description: Accès refusé
        "404":
          description: Robot introuvable
    delete:
      summary: Supprime un robot
      tags:
        - Robots
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Robot supprimé
        "403":
          description: Accès refusé
        "404":
          description: Robot introuvable
  /roles:
    get:
      summary: Liste tous les rôles
      tags:
        - Roles
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des rôles
        "403":
          description: Permission refusée
    post:
      summary: Crée un nouveau rôle
      tags:
        - Roles
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - permissions
              properties:
                name:
                  type: string
                  description: Nom du rôle
                permissions:
                  type: array
                  description: Liste des permissions associées
                  items:
                    type: string
      responses:
        "201":
          description: Rôle créé
        "400":
          description: Données invalides ou rôle existant
        "403":
          description: Permission refusée
  /roles/{id}:
    get:
      summary: Récupère un rôle par ID
      tags:
        - Roles
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rôle trouvé
        "403":
          description: Permission refusée
        "404":
          description: Rôle introuvable
    patch:
      summary: Met à jour un rôle
      tags:
        - Roles
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nouveau nom du rôle
                permissions:
                  type: array
                  description: Nouvelle liste des permissions
                  items:
                    type: string
      responses:
        "200":
          description: Rôle mis à jour
        "400":
          description: Données manquantes
        "403":
          description: Permission refusée
        "404":
          description: Rôle introuvable
    delete:
      summary: Supprime un rôle
      tags:
        - Roles
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Rôle supprimé
        "400":
          description: Rôle utilisé par des utilisateurs
        "403":
          description: Permission refusée
        "404":
          description: Rôle introuvable
...